requirements.txt
flask
python-telegram-bot==20.7
anthropic
supabase
requests